+++ RELEASES.md
@@ -2,6 +2,12 @@
 
 ShimmerRecyclerView is an custom RecyclerView library based on Facebook's [Shimmer](https://github.com/facebook/shimmer-android) effect for Android library
 
+### v0.3.0
+Now you can set xml attributes for shimmer layout reference and number of shimmer.
+
+- Use `shimmer_recycler_layout` xml attribute to set shimmer layout reference.
+- Use `shimmer_recycler_item_count` xml attribute to set number of shimmer items.
+
 ### v0.2.0
 
 Bug fixes and minor changes.